requirements.txt
Django==5.2.1
google-generativeai==0.8.0
Pillow==10.0.0
python-decouple==3.8
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
